refactor(backend): hoist common code to module
