+++ apps/backend/src/entities/seen.rs
@@ -5,16 +5,7 @@ use chrono::NaiveDate;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, SimpleObject)]
-pub struct SeenSeasonExtraInformation {
-    pub season: i32,
-    pub episode: i32,
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, FromJsonQueryResult)]
-pub enum SeenExtraInformation {
-    Show(SeenSeasonExtraInformation),
-}
+use super::utils::{SeenExtraInformation, SeenSeasonExtraInformation};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize, SimpleObject)]
 #[graphql(name = "Seen")]
