+++ apps/backend/src/media/resolver.rs
@@ -17,8 +17,9 @@ use crate::{
             AudioBook, Book, Creator, Genre, Metadata, MetadataImage, Movie, Seen, Show,
             UserToMetadata, VideoGame,
         },
-        seen::{self, SeenExtraInformation, SeenSeasonExtraInformation},
-        show, user_to_metadata, video_game,
+        seen, show, user_to_metadata,
+        utils::{SeenExtraInformation, SeenSeasonExtraInformation},
+        video_game,
     },
     graphql::IdObject,
     migrator::{MetadataImageLot, MetadataLot},
